\docType{data}
\name{fastsegData}
\alias{fastsegData}
\title{Example data set for fastseg}
\format{A simple vector with a copy number call as produced by the cn.farms algorithm.}
\description{
  The data is a small subset of copy number calls which
  were produced by the cn.farms algorithm from an
  Affymetrix SNP microarray experiment of a HapMap sample.
}
\references{
  \url{http://nar.oxfordjournals.org/content/early/2011/04/12/nar.gkr197.abstract}
  Clevert et al., cn.FARMS: a latent variable model to
  detect copy number variations in microarray data with a
  low false discovery rate, NAR, 2011
}
\keyword{data}
\keyword{datasets}

